Fixes all-day events disappearing before the end of the day.
This happens only in timezones with a negative offset relative to UTC
(that is, in the (actual) western hemisphere). All-day events are
returned with UTC time, so an all-day event on 14/08 has times
'14/08 00:00:00 +00'-'15/08 00:00:00 +00'. If the user is on UTC-3,
the end time is equivalent to '14/08 21:00:00 -03'.

(There is code that adjusts those times for the user's timezone.)

But if it is past 21:00 on 14/08 for an user in UTC-3, the end of
the event is before the start time used to query the calendar, and
the all-day event is not returned.

The fix was to adjust the start time of the query by the time zone
offset, so that all-day events that have "finished" are returned.

This also brings regular events that are already finished, but there's
code that manually filters these events if they don't fall inside
the desired range of events to show.

This does not affect users in time zones with positive offsets because
since according to the times returned by the calendar provider they
start and end the correct local time, they are returned by the query.