Fix "Fill all day events" and events spanning more than one day.
When "Fill all day events" is unchecked and there is a regular
(that is, not all-day) event that starts in one day and ends in
another one, only the entry in the starting day was displayed.

The behaviour has been changed so that the entry in the end day is
also displayed.

For all day events, only an entry for the starting day is displayed.