Removed MultidayAllDayLostEventTest.
It's not testing anything specific; it's just a general integration
test for the split of multi-day all day events (and quite
repetitive). Most of it should be already be covered; or will be
with a planned integration test that hopefully includes most of
the situations that the widget has to deal with.