Change SamsungTasksProvider to consider start date.
Even though it's not possible to set a start date on the Samsung
Calendar app, if the tasks are synced from a remote account the start
date is fetched (at least for Exchange accounts). Since there will
be a good treatment for start dates (which includes using the due
date if there is no start date), it makes sense to usem start
dates also for Samsung Tasks, if available.